third_party/WebKit/LayoutTests/platform/win/svg/animations/animationElementTiming-use-counter-expected.txt

This is a testharness.js-based test.
PASS getStartTime is use counted 
Harness: the test ran to completion.
